+++ WebApi/Controllers/BettingController.cs
@@ -18,11 +18,11 @@ namespace WebApi.Controllers
         }
         /// <summary>
         /// Returns the log loss of the first number of games passed in, for a given year
-        /// Example Call: http://localhost:32616/api/betting/GetBettingResultOfYearByNumberOfGames?year=2021&numberofgames=10
+        /// Example Call: http://localhost:32616/api/betting/GetBettingResultOfYearByNumberOfGames?year=2021&numberofgames=10&betAmount=100
         /// </summary>
         /// <returns>List of ModelLogLosses</returns>
         [HttpGet]
-        public async Task<IResult> GetBettingResultOfYearByNumberOfGames(int year, int numberOfGames, decimal betAmount)
+        public async Task<IResult> GetBettingResultOfYearByNumberOfGames(int year, int numberOfGames, double betAmount)
         {
             var betResults = await _bettingCalculator.CalculateBetOutcomes(year, numberOfGames, betAmount);
             return Results.Ok(betResults);
